pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cross-domain-recommendation"
version = "0.1.0"
description = "A modern cross-domain recommendation system"
authors = [{name = "AI Assistant", email = "ai@example.com"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.10.0",
    "pyyaml>=6.0",
    "implicit>=0.7.0",
    "lightfm>=1.17",
    "surprise>=1.1.3",
    "sentence-transformers>=2.2.2",
    "faiss-cpu>=1.7.4",
    "torch>=2.0.0",
    "tensorflow>=2.13.0",
    "tensorflow-recommenders>=0.7.3",
    "matplotlib>=3.7.0",
    "plotly>=5.15.0",
    "streamlit>=1.25.0",
    "umap-learn>=0.5.3",
    "pytest>=7.4.0",
    "black>=23.7.0",
    "ruff>=0.0.280",
    "pre-commit>=3.3.0",
    "tqdm>=4.65.0",
    "mlflow>=2.5.0",
    "wandb>=0.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "ruff>=0.0.280",
    "pre-commit>=3.3.0",
]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.ruff]
target-version = "py310"
line-length = 88
select = ["E", "F", "W", "C90", "I", "N", "UP", "YTT", "S", "BLE", "FBT", "B", "A", "COM", "C4", "DTZ", "T10", "EM", "EXE", "FA", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SLOT", "SIM", "TID", "TCH", "INT", "ARG", "PTH", "TD", "FIX", "ERA", "PD", "PGH", "PL", "TRY", "FLY", "NPY", "AIR", "PERF", "FURB", "LOG", "RUF"]
ignore = ["S101", "PLR0913", "PLR0912", "PLR0915"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
